Fixed price type

+++ pkg/model/resource.go
@@ -14,8 +14,8 @@ type Resource struct {
 }
 
 type Plan struct {
-	Name  string `json:"name"`
-	Price int64  `json:"price"`
+	Name  string  `json:"name"`
+	Price float64 `json:"price"`
 }
 
 type PatchNote struct {
